requirements.txt
flask~=3.0.2
Flask-RESTful~=0.3.10
marshmallow~=3.20.2
pytest==8.0.0
requests~=2.31.0
python-dotenv~=1.0.1
redis~=5.0.1